+++ CHANGELOG.md
@@ -19,6 +19,8 @@
   - "Biology/Birds/Seabird colony regulated areas" from NatureMap
   - "Biology/Mammals/Polar bear denning areas" from NatureMap
   - "Biology/Mammals/Musk oxen calving areas" from NatureMap
+  - "Biology/Mammals/Narwhal summer areas" from NatureMap
+  - "Biology/Mammals/Narwhal winter and spring areas" from NatureMap
 
 # v4.0.0alpha1 (2026-03-04)